[SQUASHED] Implement XOSP Blur personalization options (1/2) All credits for original work goes to serjar (Author of the very famous Xposed module: Blurred System UI)
Credits goes also to: Aaahh  for the initial implementation of it in MarshROM and
continued by me after in XOSP

Original repository of the Xposed Module: https://github.com/serajr/Blurred-System-UI

Available options:

- Blurred expanded StatusBar
- Blurred StausBar header (Deprecated on N)
- Translucent Notifications
- Blurred QS in Notification Header expanded
- Ability to change Blur Radius and Blur scale

Squash all my previous commits from Marshmallow and adapt everything to work on N:

2f006d0
d4309fa
543d7f5
738338b

//TO-DO

- Adapt and finish Recents Blur code for Nougat
- Make Translucent Notifications work (Because right now it's completely wrong, so it will
be cut off from Settings)

 Blur: Implement Recents Blur (2/2)

- Implement basic implementation of serajr Recents blur
- Added the ability to change radius and scale on user side

XOSP-Project/android_packages_apps_Settings@602d1fa

XOSP-Project/android_packages_apps_Settings@e07ad34

 Blur: Color Filters implementation (2/2)

- Since a part of the code was already available in the Blur base code, let's fix the
remaining stuff and use it.

How does it works?

- It uses the lightness level of the background in use to determine which color to apply
for the Blur bckground. So for example the background is a more lighter one a different
color will be used and so on...
- If this cannot be determined, a third type of color filter will be used which is like a
sort of mixed combination of the other two  ones, which kinda good looking.

There are three color filters:

- Light color filter (Used for Light backgrounds)
- Dark color filter (Used for Dark backgrounds)
- Mixed color filter (Used for Dark/light backgrounds)

- I made the colors to be applied simultaneously for both Blur backgrounds: StatusBar and
RecentApps ones.
- I took this decision because in my opinion there's no reason to make three more
different ColorPicker preferences just for Recents, if there will be requests on that it
will be implemented. But for now we stick to this.

Signed-off-by: Calin <nilac8991@gmail.com>